gpolyencode in separate file
